change to steam.py
